+++ src/sentry_core.c
@@ -194,6 +194,9 @@ sentry_init(sentry_options_t *options)
             "the provided DSN \"%s\" is not valid", raw_dsn ? raw_dsn : "");
     }
 
+    sentry__run_load_installation_id(options->run, options->database_path,
+        options->dsn ? options->dsn->public_key : NULL);
+
     if (transport) {
         if (sentry__transport_startup(transport, options) != 0) {
 #ifdef SENTRY_PLATFORM_NX
@@ -881,7 +884,8 @@ sentry_set_user(sentry_value_t user)
     if (!sentry_value_is_null(user)) {
         sentry_options_t *options = sentry__options_lock();
         if (options && options->session) {
-            sentry__session_sync_user(options->session, user);
+            sentry__session_sync_user(options->session, user,
+                options->run ? options->run->installation_id : NULL);
             sentry__run_write_session(options->run, options->session);
         }
         sentry__options_unlock();