+++ src/sentry_session.h
@@ -66,6 +66,7 @@ void sentry__record_errors_on_current_session(uint32_t error_count);
 /**
  * This will update a sessions `distinct_id`, which is based on the user.
  */
-void sentry__session_sync_user(sentry_session_t *session, sentry_value_t user);
+void sentry__session_sync_user(sentry_session_t *session, sentry_value_t user,
+    const char *installation_id);
 
 #endif